requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncgui==0.6.3
asyncpg==0.31.0
attrs==25.3.0
audioop-lts==0.2.2
bcrypt==5.0.0
beautifulsoup4==4.13.5
better-profanity==0.7.0
blinker==1.9.0
cachetools==6.2.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
colorgram.py==1.2.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
discord.py==2.6.4
distro==1.9.0
docopt==0.6.2
docutils==0.22
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.119.0
filelock==3.13.1
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.6.1
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.14.0
h5py==3.15.1
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
materialyoucolor==2.0.10
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.3.2
openai==1.97.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.2
pefile==2023.2.7
pillow==11.3.0
pipreqs==0.4.13
platformdirs==4.4.0
playsound==1.2.2
pluggy==1.6.0
prettytable==3.16.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.33.0
psutil==7.1.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyNaCl==1.5.0
pyparsing==3.2.5
pypiwin32==223
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pyzmq==27.1.0
redis==7.1.0
requests==2.32.5
rich==14.2.0
ruff==0.14.9
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==68.2.2
six==1.17.0
sniffio==1.3.0
soupsieve==2.8
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webserver==0.10.3
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==2.0.0
yarg==0.1.10
yarl==1.20.1
youtube-search-python==1.6.6
yt-dlp==2025.6.30
zipp==3.23.0
